requirements.txt
torch>=2.0.0
tensordict
taichi
tqdm
hydra-core
hydra-joblib-launcher
hydra_colorlog
hydra-optuna-sweeper
welford_torch
line_profiler
tensorboard
tensorboardX
torch-tb-profiler
wandb
gpustat
opencv-python
pytorch3d
open3d
numpy
moviepy==1.0.3
imageio
imageio-ffmpeg
matplotlib
onnx
onnxruntime